+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 - Added missing translations
 ##### Changed
 - Minor improvements to the layout of the settings page
+- Changed the session id link to redirect to our Discord instead of the old wiki page
 ##### Fixed
 - Fixed a bug where the app would crash if you pressed enter while having the search bar in focus (thanks to romankrru)
 ##### Removed